+++ backend/app/Http/Controllers/Api/v1/DashboardReportController.php
@@ -35,8 +35,8 @@ class DashboardReportController extends Controller
             'overall_progress' => $this->report_service->overallProgress(null, $filter),
             'delays_by_reason' => $this->report_service->delaysByReason(null, $filter),
             'delay_reasons_impact_analysis' => $this->report_service->delayReasonsImpactAnalysis(null, $filter),
-            // 'delay_reasons_trend' => $this->report_service->delayReasonsTrend(null, "dashboard", $filter),
-            // 'delay_reasons_distribution' => $this->report_service->delayReasonsDistribution(null, $filter),
+            'delay_reasons_trend' => $this->report_service->delayReasonsTrend(null, "dashboard", $filter),
+            'delay_reasons_distribution' => $this->report_service->delayReasonsDistribution(null, $filter),
             'top_delay_reasons_comparison' => $this->report_service->topDelayReasonsComparison(null, 10, $filter),
         ];